+++ cogs/general.py
@@ -4,6 +4,8 @@ from discord import app_commands
 from discord.ext import commands
 from discord.ext.commands import Context
 
+from utils import dbmanager as db
+
 
 class General(commands.Cog, name="general"):
 	def __init__(self, bot):
@@ -15,7 +17,8 @@ class General(commands.Cog, name="general"):
 	)
 	async def help_cmd(self, context: Context) -> None:
 		embed = discord.Embed(
-			title="mcbot", description="List of available commands:", color=0x9C84EF
+			title="mcbot", description="List of available commands:",
+			color=discord.Color.from_str(await db.get_setting("embed-color")),
 		)
 		for i in self.bot.cogs:
 			cog = self.bot.get_cog(i.lower())
@@ -38,7 +41,9 @@ class General(commands.Cog, name="general"):
 	)
 	async def info_cmd(self, context: Context) -> None:
 		embed = discord.Embed(
-			title="mcbot", description="Version 0.2.0", color=0x9C84EF
+			title="mcbot",
+			description="Version 0.3.0",
+			color=discord.Color.from_str(await db.get_setting("embed-color")),
 		)
 		await context.send(embed=embed)