+++ cogs/owner.py
@@ -4,6 +4,8 @@ from discord import app_commands
 from discord.ext import commands
 from discord.ext.commands import Context
 
+from utils import dbmanager as db
+
 
 class Owner(commands.Cog, name="owner"):
 	def __init__(self, bot):
@@ -20,7 +22,7 @@ class Owner(commands.Cog, name="owner"):
 			await context.bot.tree.sync()
 			embed = discord.Embed(
 				description="Slash commands have been globally synchronized",
-				color=0x9C84EF,
+				color=discord.Color.from_str(await db.get_setting("embed-color")),
 			)
 			await context.send(embed=embed)
 			return
@@ -29,12 +31,13 @@ class Owner(commands.Cog, name="owner"):
 			await context.bot.tree.sync(guild=context.guild)
 			embed = discord.Embed(
 				description="Slash commands have been synchronized in this guild",
-				color=0x9C84EF,
+				color=discord.Color.from_str(await db.get_setting("embed-color")),
 			)
 			await context.send(embed=embed)
 			return
 		embed = discord.Embed(
-			description="The scope must be `global` or `guild`", color=0xE02B2B
+			description="The scope must be `global` or `guild`",
+			color=discord.Color.from_str(await db.get_setting("embed-error-color"))
 		)
 		await context.send(embed=embed)
 
@@ -50,7 +53,7 @@ class Owner(commands.Cog, name="owner"):
 			await context.bot.tree.sync()
 			embed = discord.Embed(
 				description="Slash commands have been globally unsynchronized",
-				color=0x9C84EF,
+				color=discord.Color.from_str(await db.get_setting("embed-color")),
 			)
 			await context.send(embed=embed)
 			return
@@ -59,13 +62,47 @@ class Owner(commands.Cog, name="owner"):
 			await context.bot.tree.sync(guild=context.guild)
 			embed = discord.Embed(
 				description="Slash commands have been unsynchronized in this guild",
-				color=0x9C84EF,
+				color=discord.Color.from_str(await db.get_setting("embed-color")),
 			)
 			await context.send(embed=embed)
 			return
 		embed = discord.Embed(
-			description="The scope must be `global` or `guild`", color=0xE02B2B
+			description="The scope must be `global` or `guild`",
+			color=discord.Color.from_str(await db.get_setting("embed-error-color"))
 		)
+		await context.send(embed=embed)
+
+	@commands.hybrid_command(
+		name="set",
+		description="Change a setting",
+	)
+	@app_commands.describe(key="Key", value="Value")
+	@commands.is_owner()
+	async def set_cmd(self, context: Context, key: str, value: str) -> None:
+		await db.set_setting(key, value)
+		embed = discord.Embed(
+			description=f"{key}:{value}",
+			color=discord.Color.from_str(await db.get_setting("embed-color")),
+		)
+		await context.send(embed=embed)
+
+	@commands.hybrid_command(
+		name="list",
+		description="List all settings",
+	)
+	@commands.is_owner()
+	async def list_cmd(self, context: Context) -> None:
+		embed = discord.Embed(
+			title="mcbot",
+			description="List of all settings:",
+			color=discord.Color.from_str(await db.get_setting("embed-color")),
+		)
+		for i in await db.list_settings():
+			embed.add_field(
+				name = i[0],
+				value = i[1],
+				inline = False
+			)
 		await context.send(embed=embed)
 
 async def setup(bot):