+++ database/default.sql
@@ -1,0 +1,7 @@
+-- Default settings
+INSERT INTO setting
+	( setting_key, setting_value )
+VALUES
+	("db-version", "1700331495_setting"),
+	("embed-color", "#40a45c"),
+	("embed-error-color", "#e02b2b")